requirements.txt
fastapi
uvicorn
requests
mcp
python-dotenv